requirements.txt
boto3==1.9.148
botocore==1.12.148
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-cors-headers==3.0.1
django-environ==0.4.5
django-filter==2.1.0
djangorestframework==3.9.3
djangorestframework-simplejwt==4.3.0
docutils==0.14
ecdsa==0.13.2
envs==1.3
future==0.17.1
gunicorn==19.9.0
idna==2.8
jmespath==0.9.4
Markdown==3.1
Pillow==6.0.0
psycopg2==2.8.2
pycryptodome==3.3.1
PyJWT==1.7.1
python-dateutil==2.8.0
python-jose-cryptodome==1.3.2
pytz==2019.1
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
stripe==2.28.1
urllib3==1.24.3
virtualenv==16.6.1
virtualenvwrapper-win==1.2.5
warrant==0.6.1
